ex037 with typecast

+++ ex037.c
@@ -8,7 +8,7 @@ int main()
     scanf("%i", &number);
 
     doub = number * 2;
-    thirdPart = number * 1 / 3;
+    thirdPart = number * (float)1 / 3;
 
     printf("Your number: %i\nIts double: %i\nIts third part: %.2f\n", number, doub, thirdPart);
